web app rest api bugs fixed

+++ src/main/java/com/shop_center/hyper_market_ali/customer/CustomerService.java
@@ -3,8 +3,9 @@ package com.shop_center.hyper_market_ali.customer;
 import java.util.List;
 import java.util.Optional;
 
-// import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import jakarta.transaction.Transactional;
 
 @Service
 public class CustomerService {
@@ -24,11 +25,34 @@ public class CustomerService {
         customerRepository.save(customer);
     }
 
+    @Transactional
+    public void updateCustomer(Integer customerId, String name, String email, Integer age) {
+
+    }
+
     public void deleteCustomerById(Integer id) {
+        boolean exists = customerRepository.existsById(id);
+        if (!exists) {
+            throw new IllegalStateException("customer with id " + id + " doesn't exist!");
+        }
+
         customerRepository.deleteById(id);
     }
 
     public Optional<Customer> findCustomerById(Integer id) {
-        return customerRepository.findById(id);
+        Optional<Customer> customerOptional = customerRepository.findById(id);
+        if (!customerOptional.isPresent()) {
+            throw new IllegalStateException("customer with id " + id + " doesn't exist!");
+        }
+
+        return customerOptional;
+    }
+
+    public void addCustomer(Customer customer) {
+        Optional<Customer> customerOptional = customerRepository.findCustomerByEmail(customer.getEmail());
+        if (customerOptional.isPresent()) {
+            throw new IllegalStateException("email taken");
+        }
+        customerRepository.save(customer);
     }
 }
